Recorrect the max domain extrinsic weight and minor refactoring
